Fix package of classes in editor-commons
