+++ editor-commons/src/com/mbrlabs/mundus/editorcommons/EventListener.java
@@ -1,0 +1,4 @@
+package com.mbrlabs.mundus.editorcommons;
+
+public interface EventListener {
+}
