--- editor-commons/src/main/com/mbrlabs/mundus/editorcommons/EventListener.java
@@ -1,4 +1,0 @@
-package main.com.mbrlabs.mundus.editorcommons;
-
-public interface EventListener {
-}
