+++ editor/src/main/com/mbrlabs/mundus/editor/core/helperlines/HelperLines.kt
@@ -22,10 +22,10 @@ import com.mbrlabs.mundus.commons.scene3d.components.TerrainComponent
 import com.badlogic.gdx.utils.Array
 import com.badlogic.gdx.utils.Disposable
 import com.mbrlabs.mundus.commons.scene3d.components.Component
-import main.com.mbrlabs.mundus.editorcommons.events.GameObjectModifiedEvent
-import main.com.mbrlabs.mundus.editorcommons.events.TerrainAddedEvent
-import main.com.mbrlabs.mundus.editorcommons.events.TerrainRemovedEvent
-import main.com.mbrlabs.mundus.editorcommons.events.TerrainVerticesChangedEvent
+import com.mbrlabs.mundus.editorcommons.events.GameObjectModifiedEvent
+import com.mbrlabs.mundus.editorcommons.events.TerrainAddedEvent
+import com.mbrlabs.mundus.editorcommons.events.TerrainRemovedEvent
+import com.mbrlabs.mundus.editorcommons.events.TerrainVerticesChangedEvent
 
 class HelperLines : TerrainVerticesChangedEvent.TerrainVerticesChangedEventListener,
         TerrainAddedEvent.TerrainAddedEventListener,
