+++ editor/src/main/com/mbrlabs/mundus/editor/events/AssetImportEvent.kt
@@ -17,7 +17,7 @@
 package com.mbrlabs.mundus.editor.events
 
 import com.mbrlabs.mundus.commons.assets.Asset
-import main.com.mbrlabs.mundus.editorcommons.Subscribe
+import com.mbrlabs.mundus.editorcommons.Subscribe
 
 /**
  * @author Marcus Brummer
