+++ editor/src/main/com/mbrlabs/mundus/editor/events/FilesDroppedEvent.kt
@@ -1,6 +1,6 @@
 package com.mbrlabs.mundus.editor.events
 
-import main.com.mbrlabs.mundus.editorcommons.Subscribe
+import com.mbrlabs.mundus.editorcommons.Subscribe
 
 /**
  * Event which is posted when files are dropped from an operating system into the mundus window.
