+++ editor/src/main/com/mbrlabs/mundus/editor/events/GlobalBrushSettingsChangedEvent.kt
@@ -16,7 +16,7 @@
 
 package com.mbrlabs.mundus.editor.events
 
-import main.com.mbrlabs.mundus.editorcommons.Subscribe
+import com.mbrlabs.mundus.editorcommons.Subscribe
 
 /**
  * Indicates a change in the static settings each TerrainBrush shares (strength,
