+++ editor/src/main/com/mbrlabs/mundus/editor/events/SceneChangedEvent.kt
@@ -16,7 +16,7 @@
 
 package com.mbrlabs.mundus.editor.events
 
-import main.com.mbrlabs.mundus.editorcommons.Subscribe
+import com.mbrlabs.mundus.editorcommons.Subscribe
 
 /**
  * @author Marcus Brummer
