+++ editor/src/main/com/mbrlabs/mundus/editor/history/commands/DeleteCommand.kt
@@ -28,7 +28,7 @@ import com.mbrlabs.mundus.editor.events.SceneGraphChangedEvent
 import com.mbrlabs.mundus.editor.history.Command
 import com.mbrlabs.mundus.editor.ui.modules.outline.Outline
 import com.mbrlabs.mundus.editor.utils.Log
-import main.com.mbrlabs.mundus.editorcommons.events.TerrainAddedEvent
+import com.mbrlabs.mundus.editorcommons.events.TerrainAddedEvent
 
 /**
  * Delete command for game objects Deletion will update sceneGraph and outline
