+++ editor/src/main/com/mbrlabs/mundus/editor/terrain/PerlinNoiseGenerator.java
@@ -21,7 +21,7 @@ import java.util.Random;
 import com.badlogic.gdx.math.Interpolation;
 import com.mbrlabs.mundus.commons.scene3d.components.TerrainComponent;
 import com.mbrlabs.mundus.editor.Mundus;
-import main.com.mbrlabs.mundus.editorcommons.events.TerrainVerticesChangedEvent;
+import com.mbrlabs.mundus.editorcommons.events.TerrainVerticesChangedEvent;
 
 /**
  *
