+++ editor/src/main/com/mbrlabs/mundus/editor/tools/TransformTool.java
@@ -21,7 +21,7 @@ import com.mbrlabs.mundus.editor.history.CommandHistory;
 import com.mbrlabs.mundus.editor.preferences.MundusPreferencesManager;
 import com.mbrlabs.mundus.editor.tools.picker.GameObjectPicker;
 import com.mbrlabs.mundus.editor.tools.picker.ToolHandlePicker;
-import main.com.mbrlabs.mundus.editorcommons.events.GameObjectModifiedEvent;
+import com.mbrlabs.mundus.editorcommons.events.GameObjectModifiedEvent;
 
 /**
  * @author Marcus Brummer
