+++ editor/src/main/com/mbrlabs/mundus/editor/tools/brushes/TerrainBrush.java
@@ -42,7 +42,7 @@ import com.mbrlabs.mundus.editor.history.commands.TerrainPaintCommand;
 import com.mbrlabs.mundus.editor.shader.EditorPBRTerrainShader;
 import com.mbrlabs.mundus.editor.tools.Tool;
 import com.mbrlabs.mundus.editor.ui.UI;
-import main.com.mbrlabs.mundus.editorcommons.events.TerrainVerticesChangedEvent;
+import com.mbrlabs.mundus.editorcommons.events.TerrainVerticesChangedEvent;
 
 /**
  * A Terrain Brush can modify the terrainAsset in various ways (BrushMode).
