+++ editor/src/test/java/com/mbrlabs/mundus/editor/events/EventBusTest.java
@@ -1,6 +1,6 @@
 package com.mbrlabs.mundus.editor.events;
 
-import main.com.mbrlabs.mundus.editorcommons.Subscribe;
+import com.mbrlabs.mundus.editorcommons.Subscribe;
 import org.junit.Assert;
 import org.junit.Test;
